% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{available_case_studies}
\alias{available_case_studies}
\title{List built-in case study names}
\usage{
available_case_studies()
}
\value{
Character vector of built-in case study names.
}
\description{
List built-in case study names
}
