% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{case_study_catalog}
\alias{case_study_catalog}
\title{Case-study catalog for \code{heteff}}
\usage{
case_study_catalog()
}
\value{
A data frame describing built-in case studies.
}
\description{
Case-study catalog for \code{heteff}
}
