% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{heteff_fit_methods}
\alias{heteff_fit_methods}
\alias{print.heteff_fit}
\alias{as.data.frame.heteff_fit}
\alias{plot.heteff_fit}
\title{Common Methods for \code{heteff_fit} Objects}
\usage{
\method{print}{heteff_fit}(x, ...)

\method{as.data.frame}{heteff_fit}(x, ...)

\method{plot}{heteff_fit}(x, ...)
}
\arguments{
\item{x}{A \code{heteff_fit} object.}

\item{...}{Not used.}
}
\description{
Common Methods for \code{heteff_fit} Objects
}
