% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.heteff_fit}
\alias{plot.heteff_fit}
\title{Plot a default subgroup-effect summary from a \code{heteff_fit}}
\usage{
\method{plot}{heteff_fit}(x, ...)
}
\arguments{
\item{x}{A \code{heteff_fit} object.}

\item{...}{Passed to \code{\link[=plot_subgroup_effects]{plot_subgroup_effects()}}.}
}
\value{
Draws a subgroup summary plot.
}
\description{
Plot a default subgroup-effect summary from a \code{heteff_fit}
}
