% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/explain.R
\name{plot_causal_tree_explorer}
\alias{plot_causal_tree_explorer}
\title{Plot a causalTree explorer result}
\usage{
plot_causal_tree_explorer(x, main = "Causal tree explorer", ...)
}
\arguments{
\item{x}{Output from \code{fit_causal_tree_explorer()}.}

\item{main}{Plot title.}

\item{...}{Reserved for future use.}
}
\value{
Draws a tree plot.
}
\description{
Plot a causalTree explorer result
}
