% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_effect_tree}
\alias{plot_effect_tree}
\title{Plot an explanation tree with a publication-style layout}
\usage{
plot_effect_tree(fit, main = "Explanation tree")
}
\arguments{
\item{fit}{A \code{heteff_fit} object.}

\item{main}{Plot title.}
}
\value{
Draws a tree plot.
}
\description{
Plot an explanation tree with a publication-style layout
}
