% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_observational_dag}
\alias{plot_observational_dag}
\title{Plot the observational causal DAG}
\usage{
plot_observational_dag()
}
\value{
A \code{ggplot2} DAG plot.
}
\description{
Plot the observational causal DAG
}
