% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_variable_importance}
\alias{plot_variable_importance}
\title{Plot ranked variable importance}
\usage{
plot_variable_importance(fit, top_n = 12)
}
\arguments{
\item{fit}{A \code{heteff_fit} object.}

\item{top_n}{Number of top covariates to show.}
}
\value{
A \code{ggplot2} object when available; otherwise draws a base plot.
}
\description{
Plot ranked variable importance
}
