% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.heteff_fit}
\alias{print.heteff_fit}
\title{Print a \code{heteff_fit} summary}
\usage{
\method{print}{heteff_fit}(x, ...)
}
\arguments{
\item{x}{A \code{heteff_fit} object.}

\item{...}{Unused.}
}
\value{
The input object, invisibly.
}
\description{
Print a \code{heteff_fit} summary
}
