% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cases.R
\name{run_case_study}
\alias{run_case_study}
\title{Run a built-in case study}
\usage{
run_case_study(case = available_case_studies(), num_trees = 500, seed = 123)
}
\arguments{
\item{case}{Study name. One of \code{available_case_studies()}.}

\item{num_trees}{Number of trees for the selected workflow.}

\item{seed}{Optional seed.}
}
\value{
A \code{heteff_fit} object.
}
\description{
Run a built-in case study
}
