% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heteff-package.R
\docType{package}
\name{heteff-package}
\alias{heteff}
\alias{heteff-package}
\title{heteff: Minimal Forest Workflows for Heterogeneous Effect Estimation}
\description{
\code{heteff} provides compact wrappers around selected estimators from
\href{https://grf-labs.github.io/grf/}{grf}, focused on heterogeneous effects for
observational, survival, and instrumental-variable settings.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/dai540/heteff}
  \item \url{https://dai540.github.io/heteff/}
  \item Report bugs at \url{https://github.com/dai540/heteff/issues}
}

}
\author{
\strong{Maintainer}: Dai \email{daik54015@gmail.com}

}
\keyword{internal}
